Added Disassembler code and style fixes
Fixed the structure of the extractOpCode function to align with agreed on conventions and tab indented the switch case menu to align with C/C++ conventions. 
Added the 2nd third of the cases, provided documentation and added documentation for who was assigned what part for the sake of showing the grader who did what.